+++ src/Panel.tsx
@@ -36,7 +36,8 @@ export const Panel = ({ active, api }: PanelProps) => {
 
   const [gitInfo] = useAddonState<GitInfoPayload>(GIT_INFO);
   const [gitInfoError] = useAddonState<Error>(GIT_INFO_ERROR);
-  const [localBuildProgress] = useAddonState<LocalBuildProgress>(LOCAL_BUILD_PROGRESS);
+  const [localBuildProgress, setLocalBuildProgress] =
+    useAddonState<LocalBuildProgress>(LOCAL_BUILD_PROGRESS);
   const [, setOutdated] = useAddonState<boolean>(IS_OUTDATED);
   const emit = useChannel({});
 
@@ -138,18 +139,19 @@ export const Panel = ({ active, api }: PanelProps) => {
     );
   }
 
-  const localBuildIsRightBranch = gitInfo && gitInfo.branch === localBuildProgress?.branch;
+  const localBuildIsRightBranch = gitInfo.branch === localBuildProgress?.branch;
   return (
     <Provider key={PANEL_ID} value={client}>
       <Sections hidden={!active}>
         <VisualTests
-          projectId={projectId}
-          gitInfo={gitInfo}
+          dismissBuildError={() => setLocalBuildProgress(undefined)}
           localBuildProgress={localBuildIsRightBranch ? localBuildProgress : undefined}
           startDevBuild={() => emit(START_BUILD)}
           setAccessToken={setAccessToken}
           setOutdated={setOutdated}
           updateBuildStatus={updateBuildStatus}
+          projectId={projectId}
+          gitInfo={gitInfo}
           storyId={storyId}
         />
       </Sections>