+++ src/components/SidebarTopButton.stories.tsx
@@ -1,7 +1,8 @@
 import { action } from "@storybook/addon-actions";
-import { expect } from "@storybook/jest";
 import type { Meta, StoryObj } from "@storybook/react";
 import { screen, userEvent, within } from "@storybook/testing-library";
+import isChromatic from "chromatic/isChromatic";
+import React, { ComponentProps, useEffect, useState } from "react";
 
 import { INITIAL_BUILD_PAYLOAD } from "../buildSteps";
 import { playAll } from "../utils/playAll";
@@ -11,6 +12,7 @@ const meta = {
   component: SidebarTopButton,
   args: {
     startBuild: action("startBuild"),
+    stopBuild: action("stopBuild"),
   },
 } satisfies Meta<typeof SidebarTopButton>;
 
@@ -32,18 +34,33 @@ export const Outdated: Story = {
   },
 };
 
+const WithProgress = (props: ComponentProps<typeof SidebarTopButton>) => {
+  const [buildProgressPercentage, setProgress] = useState(20);
+  useEffect(() => {
+    if (isChromatic()) return () => {};
+    const interval = setInterval(() => {
+      setProgress((p) => (p < 100 ? p + 1 : 0));
+    }, 300);
+    return () => clearInterval(interval);
+  }, []);
+  return (
+    <SidebarTopButton
+      {...props}
+      isRunning
+      localBuildProgress={{
+        ...INITIAL_BUILD_PAYLOAD,
+        buildProgressPercentage,
+        currentStep: "build",
+      }}
+    />
+  );
+};
+
 export const IsRunning: Story = {
-  args: {
-    isRunning: true,
-    localBuildProgress: {
-      ...INITIAL_BUILD_PAYLOAD,
-      buildProgressPercentage: 40,
-      currentStep: "build",
-    },
-  },
+  render: WithProgress,
   play: playAll(async ({ canvasElement }) => {
     const canvas = within(canvasElement);
-    const button = await canvas.findByRole("button", { name: "Run tests" });
+    const button = await canvas.findByRole("button", { name: "Stop tests" });
     // Wait one second just to ensure the screen has proper focus
     await new Promise((r) => setTimeout(r, 1000));
     await userEvent.hover(button);