+++ src/screens/VisualTests/VisualTests.tsx
@@ -31,22 +31,24 @@ const createEmptyStoryStatusUpdate = (state: API_StatusState) => {
 interface VisualTestsProps {
   selectedBuildInfo?: SelectedBuildInfo;
   setSelectedBuildInfo: ReturnType<typeof useState<SelectedBuildInfo>>[1];
-  projectId: string;
-  gitInfo: Pick<
-    GitInfoPayload,
-    "branch" | "slug" | "userEmailHash" | "commit" | "committedAt" | "uncommittedHash"
-  >;
+  dismissBuildError: () => void;
   localBuildProgress?: LocalBuildProgress;
   startDevBuild: () => void;
   setAccessToken: (accessToken: string | null) => void;
   setOutdated: (isOutdated: boolean) => void;
   updateBuildStatus: UpdateStatusFunction;
+  projectId: string;
+  gitInfo: Pick<
+    GitInfoPayload,
+    "branch" | "slug" | "userEmailHash" | "commit" | "committedAt" | "uncommittedHash"
+  >;
   storyId: string;
 }
 
 export const VisualTestsWithoutSelectedBuildId = ({
   selectedBuildInfo,
   setSelectedBuildInfo,
+  dismissBuildError,
   localBuildProgress,
   startDevBuild,
   setAccessToken,
@@ -230,6 +232,7 @@ export const VisualTestsWithoutSelectedBuildId = ({
     <BuildResults
       {...{
         branch: gitInfo.branch,
+        dismissBuildError,
         localBuildProgress,
         ...(lastBuildOnBranch && { lastBuildOnBranch }),
         lastBuildOnBranchCompletedStory,