+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# v0.0.105 (Tue Oct 10 2023)
+
+#### 🐛 Bug Fix
+
+- Allow build cancelation [#123](https://github.com/chromaui/addon-visual-tests/pull/123) ([@ghengeveld](https://github.com/ghengeveld) [@chromatic-support](https://github.com/chromatic-support))
+
+#### Authors: 2
+
+- [@chromatic-support](https://github.com/chromatic-support)
+- Gert Hengeveld ([@ghengeveld](https://github.com/ghengeveld))
+
+---
+
+# v0.0.104 (Tue Oct 10 2023)
+
+#### 🐛 Bug Fix
+
+- Fix z-index to prevent weird rendering bug with Storybook panel [#129](https://github.com/chromaui/addon-visual-tests/pull/129) ([@chromatic-support](https://github.com/chromatic-support) [@ghengeveld](https://github.com/ghengeveld))
+
+#### Authors: 2
+
+- [@chromatic-support](https://github.com/chromatic-support)
+- Gert Hengeveld ([@ghengeveld](https://github.com/ghengeveld))
+
+---
+
 # v0.0.103 (Sun Oct 08 2023)
 
 #### 🐛 Bug Fix